Adjust passport option, fixes certain flash messages not showing
Any flash messages related to connecting an additional account to an existing user account, and also after a successful login, were not showing because the session was regenerated and flash messages were removed from the session before being shown. Setting keepSessionInfo to true makes sure that these messages do not get removed when the session is regenerated.